pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.3.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "maria", email = "mariaeduarda.gerotti@gmail.com"},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
